+++ src/components/index.ts
@@ -2,8 +2,8 @@
 export { Reveal } from "./Reveal";
 export { Counter } from "./Counter";
 export { Masthead } from "./Masthead";
+export { ReadingProgress } from "./ReadingProgress";
 export { ReceiptSlip } from "./ReceiptSlip";
-export { SectionFallback } from "./SectionFallback";
 export { NightScene } from "./art/NightScene";
 export { RadialClock } from "./viz/RadialClock";
 export { ConstellationGraph } from "./viz/ConstellationGraph";